pyproject.toml
[tool.poetry]
name = "smart-store-project"
version = "0.1.0"
description = ""
authors = ["kimdooh2019 <choiinhee2020@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "<3.0"
notebook = "^6.1.4"
django-extensions = "^3.0.9"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^5.0.0"}
djangorestframework = "^3.11.1"
requests = "^2.24.0"
django-rest-auth = "^0.9.5"
pandas = "^1.1.2"
powernad = "^0.8.0"
jsonpickle = "^1.4.1"
django-cors-headers = "^3.5.0"
xlrd = "^1.2.0"
supervisor = "^4.2.1"
gunicorn = "^20.0.4"
awscli = "^1.18.178"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
